Fix DAO reference issue in onActivityResult.
If the Activity needs to be recreated in when the activity returns some
results to the parent activity. All the local variables will lost and
needs to be re-created asynchronized. But onActivityResult is called
right after onCreate and these variable will not be created immediately.
This change modify that getDbOpenHelper will be available immediately
after onCreate so onActivityResult can use it.

This change also all DAO to use getDbOpenHelper to get the instance
instead of using Activity cached one.